Tilt done

+++ dev/app/src/main/java/world/augma/ui/AR/OnRotationChangedListener.java
@@ -1,5 +1,5 @@
 package world.augma.ui.AR;
 
 public interface OnRotationChangedListener {
-    void onRotationChanged(float newRot,float inclination,int up);
+    void onRotationChanged(float newRot,int inclination, int up);
 }
